# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi Dosen",
        "description": "Dokumentasi API untuk manajemen data dosen",
        "version": "1.0"
    },
    "paths": {
        "/api/dosen": {
            "get": {
                "tags": [
                    "Dosen"
                ],
                "summary": "Ambil semua data dosen",
                "operationId": "fe689a813cdf4da967ff604ab6948700",
                "responses": {
                    "200": {
                        "description": "Daftar dosen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Dosen"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dosen"
                ],
                "summary": "Tambah dosen baru",
                "operationId": "5bb3790e30481b44aa318120de9556c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_dosen",
                                    "nidn",
                                    "email",
                                    "alamat",
                                    "program_studi",
                                    "tanggal_lahir",
                                    "jenis_kelamin",
                                    "status",
                                    "bidang_keahlian"
                                ],
                                "properties": {
                                    "nama_dosen": {
                                        "type": "string",
                                        "example": "Dr. Diksa"
                                    },
                                    "nidn": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "diksa@example.com"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 123"
                                    },
                                    "program_studi": {
                                        "type": "string",
                                        "example": "Teknik Informatika"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1980-05-10"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "L",
                                            "P"
                                        ],
                                        "example": "L"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Dosen Tetap",
                                            "Dosen Tidak Tetap"
                                        ],
                                        "example": "Dosen Tetap"
                                    },
                                    "bidang_keahlian": {
                                        "type": "string",
                                        "example": "Kecerdasan Buatan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dosen ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dosen/{id}": {
            "get": {
                "tags": [
                    "Dosen"
                ],
                "summary": "Ambil data dosen berdasarkan ID",
                "operationId": "d779c71648b70b8892c18d49dc24589a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dosen ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dosen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dosen tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dosen"
                ],
                "summary": "Update data dosen",
                "operationId": "39299cb50b2967b671b945f0051f0f4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_dosen": {
                                        "type": "string",
                                        "example": "Dr. Budi"
                                    },
                                    "nidn": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "budi@example.com"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 123"
                                    },
                                    "program_studi": {
                                        "type": "string",
                                        "example": "Teknik Elektro"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1985-03-15"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "L",
                                            "P"
                                        ],
                                        "example": "L"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Dosen Tetap",
                                            "Dosen Tidak Tetap"
                                        ],
                                        "example": "Dosen Tidak Tetap"
                                    },
                                    "bidang_keahlian": {
                                        "type": "string",
                                        "example": "Jaringan Komputer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dosen diperbarui"
                    },
                    "404": {
                        "description": "Dosen tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dosen"
                ],
                "summary": "Hapus dosen",
                "operationId": "f8a589fab34ccf23402cbd5e5d0b4a58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dosen dihapus"
                    },
                    "404": {
                        "description": "Dosen tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Dosen": {
                "title": "Dosen",
                "required": [
                    "id",
                    "nama_dosen",
                    "nidn"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_dosen": {
                        "type": "string",
                        "example": "Dr. Ahmad"
                    },
                    "nidn": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "email": {
                        "type": "string",
                        "example": "ahmad@example.com"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Merdeka No. 123"
                    },
                    "program_studi": {
                        "type": "string",
                        "example": "Teknik Informatika"
                    },
                    "tanggal_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "1980-01-01"
                    },
                    "jenis_kelamin": {
                        "type": "string",
                        "enum": [
                            "L",
                            "P"
                        ],
                        "example": "L"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Dosen Tetap",
                            "Dosen Tidak Tetap"
                        ],
                        "example": "Dosen Tetap"
                    },
                    "bidang_keahlian": {
                        "type": "string",
                        "example": "Kecerdasan Buatan"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}